+++ flutter/lib/src/event_processor/native_app_start_event_processor.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
-import 'package:sentry/sentry.dart';
-
+import '../../sentry_flutter.dart';
 import '../integrations/integrations.dart';
 import '../native/sentry_native.dart';
 
@@ -9,16 +8,32 @@ import '../native/sentry_native.dart';
 /// measurement.
 class NativeAppStartEventProcessor implements EventProcessor {
   final SentryNative _native;
+  final Hub _hub;
 
-  NativeAppStartEventProcessor(this._native);
+  NativeAppStartEventProcessor(this._native, this._hub);
 
   @override
   Future<SentryEvent?> apply(SentryEvent event, {Hint? hint}) async {
-    if (_native.didAddAppStartMeasurement || event is! SentryTransaction) {
+    // ignore: invalid_use_of_internal_member
+    final options = _hub.options;
+    if (_native.didAddAppStartMeasurement ||
+        event is! SentryTransaction ||
+        options is! SentryFlutterOptions) {
       return event;
     }
 
     final appStartInfo = await NativeAppStartIntegration.getAppStartInfo();
+
+    final appStartEnd = _native.appStartEnd;
+    if (!options.autoAppStart) {
+      if (appStartEnd != null) {
+        appStartInfo?.end = appStartEnd;
+      } else {
+        // If autoAppStart is disabled and appStartEnd is not set, we can't add app starts
+        return event;
+      }
+    }
+
     final measurement = appStartInfo?.toMeasurement();
 
     if (measurement != null) {