+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Event processor blocking transactions from being sent if `autoAppStart` is false ([#2045](https://github.com/getsentry/sentry-dart/pull/2045))
+
 ## 7.20.1
 
 ### Dependencies